clean up: merge arthswap and stellaswap
