+++ index/src/executors/dex_executors/mod.rs
@@ -1,4 +1,3 @@
 pub mod acala_dot_swap;
 pub mod acala_swap;
-pub mod arth_swap;
-pub mod moonbeam_swap;
+pub mod uniswap_based;
